requirements.txt
# Core dependencies
beautifulsoup4==4.13.3
python-dotenv==1.0.1
PyPDF2==3.0.1
requests==2.32.3
sseclient-py==1.8.0
streamlit==1.43.1
aiohttp==3.9.3
asyncio==3.4.3
markdown==3.5.2

# Optional but recommended for better performance
pytest==7.4.3  # For testing
black==23.11.0  # For code formatting

# The following dependencies are automatically installed by the above packages
# but are listed here for reference:
# certifi
# charset-normalizer
# click
# importlib-metadata
# packaging
# pillow
# protobuf
# python-dateutil
# six
# soupsieve
# typing-extensions
# urllib3
